requirements.txt
torch==2.5.1
transformers==5.3.0
sentence-transformers==5.3.0
numpy==1.26.4
pandas==3.0.1
scikit-learn==1.8.0
matplotlib==3.10.8
nltk==3.9.3
jupyter==1.0.0
tqdm==4.67.3
python-dotenv==1.0.1
